Add files via upload
This paper performs the usual calculations and table manipulations using the pandas library. As a result of the study, we obtained an index model that aggregates diverse data from seven tables. This model has two metrics - growth and percentage growth by decades. p.s. The purpose of this work was to make calculations for my graduate thesis, for reasons of political correctness, the conclusions of the academic part will not be published.